Add ticket view

+++ shared/src/commonMain/kotlin/com/mrebollob/loteria/domain/entity/Ticket.kt
@@ -3,5 +3,5 @@ package com.mrebollob.loteria.domain.entity
 data class Ticket(
     val name: String,
     val number: Int,
-    val bet: Int
+    val bet: Float
 )
